requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pillow==11.1.0
python-dotenv==1.0.1
django-filter==24.3
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
coverage==7.6.10
drf-yasg==1.21.8
stripe==11.5.0
requests==2.32.3
redis==5.2.1
celery==5.4.0